pyproject.toml
[project]
name = "aemo-electricity-prices"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "openelectricity>=0.9.2",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
]
